Move SVE functions into their own files.
